import curses library, dependencies and required tools from netbsd
in order to have a real replacement for ncurses, we need to provide
all the libraries that ncurses provide, as well as some of its tools.

on the library side, these are:
libcurses, libform, libmenu, libpanel, libterminfo (origin lib/).

on the utils side we need:
nbperf, tic, tput, tset (origin usr.bin/).
(nbperf is merely a build requirement for libterminfo).

additionally the build requires the terminfo database (origin share/),
to build an optimized hardcoded terminfo db into the libraries.

all of the files are taken from the netbsd src tree as of commit
7bcc9d8 from this git mirror: https://github.com/jsonn/src/
the last commit is from:

Author: christos <christos>
Date:   Sun Jan 31 18:57:29 2016 +0000

    PR/50729: Izumi Tsutsui: Add "SMALLPROG"-like options to disklabel(8)