fix portability issues and provide top-level GNUmakefile
the original code uses several nonportable features, including, but
not limited to:
- usage of sys/cdefs.h
- ugly BEGIN/END_DECLS macro usage in public headers
- usage of sys/queue.h (even in public headers)
- usage of funopen()
...

some source files that are built into netbsd's libc were copied
from the same revision as the initial import.

the new top-level makefile builds .a and .so with separate object
files (unless -fPIC is set in CFLAGS) and is cross-compile friendly.
for cross-compiles, nbperf is built with HOSTCC to generate some
necessary source files containing hash tables.
tic is built twice to be used for the build and to be installed on
the host.
there is a lot of room for improvement in the Makefile; but for the
moment it works sufficiently well.