term.c: disable usage of TERMCAP env var
ncurses disables it by default as well.
basically when it is used, it enables only the capabilities that existed
in the termcap age, and the terminfo db (whether compiled in or external)
will not be used. the most visible effect of this is that colors can not be
set, as this capability did not exist[0].

[0] http://www.delorie.com/gnu/docs/termcap/termcap_44.html