initscr(): don't return NULL when term is not found
adapt to the behaviour of ncurses by printing an error message and exiting
rather than returning a NULL pointer and make subsequent calls to the curses
lib crash.
this is behaviour documented by the ncurses initscr() manpage, and so there's
lots of code that doesn't check the return value, but assumes it cannot fail
(other than gracefully exiting).

steps to reproduce: TERM=foobar some-curses-app -> segfault