curses.h: add no-op implementation of mouse functions/macros
since ncurses presence is assumed by lots of linux software,
checks for availability of ncurses extensions are usually omitted.
one such extension is the mouse mode.
since netbsd curses lacks it, we implement it as no-op macros
so software using it compiles and the code dealing with it is
automatically stripped out via dead-code elimination.