make it possible to do static/dynamic builds/installs only
the all-static and install-static makefile targets can be used
to build or install only the static libs,
and all-dynamic/install-dynamic vice versa for dynamic libs only.

if one wants the programs (tset, tput, tic) to be *entirely*
static linked (i.e. not only against libterminfo but also against
libc), -static has to be added to LDFLAGS by the user, either
by putting it into config.mak or passing it to the make invocation.

closes #5