Makefile: fix issue preventing parallel builds
the explicit -lterminfo lead to ld trying to use the dynamic library
which was not yet entirely created even when TI_LINKLIB was set to use
the static one.
this caused build problems with -j8 on an octacore pc.
since the TI_LINKLIB is explicitly added to the command line, we can
just delete the gratuitous linker instructions.