keyname.c: always build keyname information
the keyname functionality is crucial to get python to work, which
(due to the lack of has_key() in our implementation) uses a
keyname()-based fallback implementation written in python.
failure to hand out the right keynames prevents python's curses module
from adding members to its _curses struct, and causes immediate error
when importing the curses module, as it tries to access those members.

the functionality was included only if the SMALL preprocessor macro
was not defined, supposedly because the existing implementation was
very bloated.
i rewrote it to use one of 3 size-optimized string tables.
the entire code + tables is now using 1.3 KB when compiled for x86_64
with -Os, which is small enough to compile it in unconditionally.