Makefile: fix logic pulling in either terminfo .a or .so for binaries
the TI_LINKLIB target specific variable, which was intended to be
overridden via "make all-dynamic" wasn't actually evaluated, and
the tset and tputs binaries always linked against the static archive.

since it's tricky to express this logic, we now make it conditional
on whether the user puts -static in his LDFLAGS, which hints that
he wants to use the static libs to compile the programs.