remove redundant __RCSID macros
the exact same info is already given in the first line of each source
file.

since the __RCSID always follows the inclusion of sys/queue.h, which
we had to change, this prevents clean application of almost any upstream patch,
so that specific hunk has to be removed.
in order not to be forced to manually update the RCSID with the value from
the first line for consistency, it's better to just get rid of it entirely.