terminfo: fix linux terminal description
midnight commander under linux terminal was heavily broken.
the root cause of this problem are wrong terminfo entries, that once
were fixed[0], but then reverted[1] by pointing the linux terminfo entry
to use the unfixed linux2.2 entry instead of the recent and fixed
linux3.0 entry.
for us this fixes it mostly but not completely, only when we use the
ascs entry from bterm, the line drawing characters are finally drawn
correctly.

changes in this commit:
- default linux to use fixed linux3.0 entry
- add improved ascs capability to linux3.0

fixes #22

[0]: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=515609
[1]: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=665959