curses.h: put the whole thing into an extern "C" block for C++
currently a C++ program that uses stdscr fails to link because
the stdscr variable is not defined with C linkage.

putting extern "C" blocks here and there as-needed is error-prone, so
we simply put the entire header inside one.
this is also what was recommended by rich felker, the author of
musl libc, and it seems to work out rather well there.