libpanel.so: Fix a build break, without an already installed version
While testing the new rc of tinycc on linux, the build breaks at libpanel/libpanel.so:
* libpanel/libpanel.so depends on libcurses/libcurses.so

* libcurses/libcurses.so depends on libterminfo/libterminfo.so
  the compiler searches for libterminfo.so, but fails.

This patch adds the missing reference to libterminfo/ during the build
to allow the compiler to find libterminfo.so.

This bug will never be visible, when the os has a libterminfo.so installed
in one of the default lib search directories.

--
Regards ... Detlef