really fix termdb lookup when new GNU screen versions are used
newer versions of GNU screen (or at least the one shipped with ubuntu)
set the TERM variable to screen.XXX, where XXX is the original TERM.

naturally, we didn't understand that we have to look after the dot
for the real terminal name in DB lookups, so this effectively broke
any terminal app, unless the user manually reset the TERM variable.

we now fix the issue one level deeper, so it affects both curses
and termcap/terminfo users, unlike the previous patch.